% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{lidar_app}
\alias{lidar_app}
\title{function to launch the lidar app}
\usage{
lidar_app(user = "guest", password = "guest", host = NULL,
  port = NULL, dbname = "lidargis")
}
\description{
function to launch the lidar app
}
