% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mainDataOutput.R
\name{mod_mainData}
\alias{mod_mainData}
\title{mod_mainData server function}
\usage{
mod_mainData(
  input,
  output,
  session,
  lang,
  app_translations,
  data_reactives,
  map_reactives,
  lidardb
)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang selected}

\item{app_translations}{dictionary}

\item{data_reactives, map_reactives}{reactives needed to retrieve the data}

\item{lidardb}{access to lidar database}
}
\description{
mod_mainData server function
}
