% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapOutput.R
\name{mod_mapOutput}
\alias{mod_mapOutput}
\alias{mod_map}
\title{mod_mapOutput and mod_map}
\usage{
mod_mapOutput(id)

mod_map(
  input,
  output,
  session,
  lang,
  app_translations,
  main_data_reactives,
  data_reactives
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang selected}

\item{app_translations}{dictionary}

\item{main_data_reactives, data_reactives}{reactives needed}
}
\description{
Shiny module to generate the map
}
