% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_saveOutput.R
\name{mod_saveOutput}
\alias{mod_saveOutput}
\alias{mod_save}
\title{mod_saveOutput and mod_save}
\usage{
mod_saveOutput(id)

mod_save(input, output, session, lang, app_translations, main_data_reactives)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang selected}

\item{app_translations}{dictionary}

\item{main_data_reactives}{reactives needed}
}
\description{
module for creating the astounding viz when click
}
