% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tableOutput.R
\name{mod_tableOutput}
\alias{mod_tableOutput}
\alias{mod_table}
\title{mod_tableOutput and mod_table}
\usage{
mod_tableOutput(id)

mod_table(input, output, session, lang, app_translations, main_data_reactives)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang selected}

\item{app_translations}{dictionary}

\item{main_data_reactives}{reactives needed}
}
\description{
Shiny module to generate the table
}
