% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tabTranslateOutput.R
\name{mod_tabTranslateOutput}
\alias{mod_tabTranslateOutput}
\alias{mod_tabTranslate}
\title{mod_tabTranslateOutput and mod_tabTranslate}
\usage{
mod_tabTranslateOutput(id)

mod_tabTranslate(input, output, session, lang, app_translations, tab_title)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang selected}

\item{app_translations}{dictionary}

\item{tab_title}{title to translate (character)}
}
\description{
A shiny module to translate tab titles
}
