% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_clip.R
\name{lidar_clip}
\alias{lidar_clip}
\title{Clip and summarise polygons}
\usage{
lidar_clip(sf, lidar_db, poly_id, lang, safe = TRUE, ...)
}
\arguments{
\item{sf}{sf object with the polygons to clip}

\item{lidar_db}{dbi connection}

\item{poly_id}{sf column name containing the polygon identificator, as character}

\item{lang}{language for the errors}

\item{safe}{logical indicating if memory and time safeguards are active}

\item{...}{not implemented}
}
\value{
a sf data frame with the variables mean (each var a column) for each polygon
  (rows) as well as the geometry column
}
\description{
lidar_clip clips polygons on to the lidar variables rasters to get the mean value
}
