% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infoUI.R
\name{mod_info}
\alias{mod_info}
\title{mod_info server function}
\usage{
mod_info(
  input,
  output,
  session,
  lang,
  app_translations,
  main_data_reactives,
  data_reactives,
  id_click
)
}
\arguments{
\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang selected}

\item{app_translations}{dictionary}

\item{main_data_reactives, data_reactives}{reactives needed}

\item{id_click}{character with the id of the shape/row clicked}
}
\description{
mod_info server function
}
